+++ tests/test_dcos_e2e/test_cluster.py
@@ -16,7 +16,7 @@ from _pytest.logging import LogCaptureFixture
 from kazoo.client import KazooClient
 from py.path import local  # pylint: disable=no-name-in-module, import-error
 
-from dcos_e2e.backends import ClusterBackend
+from dcos_e2e.base_classes import ClusterBackend
 from dcos_e2e.cluster import Cluster
 from dcos_e2e.node import Output