+++ tests/test_dcos_e2e/test_legacy.py
@@ -11,7 +11,7 @@ from pathlib import Path
 from kazoo.client import KazooClient
 from passlib.hash import sha512_crypt
 
-from dcos_e2e.backends import ClusterBackend
+from dcos_e2e.base_classes import ClusterBackend
 from dcos_e2e.cluster import Cluster
 from dcos_e2e.node import Output